Feature branch: streaming repo sync (#542)
* Scaffold xrpc-stream

* Write and test stream frames

* Write and test stream server

* Test bad stream frame parsing cases

* Proof of concept streaming on xrpc-server

* Test bad streaming endpoint

* Make frame stream to fix buffering frame bytes together

* reorg

* some lex-cli things

* better codegen for subscriptions

* scaffolding repo subscriptions

* wip

* remove repo ops

* setup notify/listen for db

* end pool to fix hanging test

* small comment in test

* basic sequencer

* some refactoring

* switch to event emitter

* reconnect on listener error

* rename notifyClient

* remove payload on channels

* pr feedback

* subscribeRepo outbox

* some cleanup

* wip

* wip

* bugfixin

* only send msgs after tx is committed

* better handle event-emitter -> generator

* max buffer size

* cleanup

* Sync-up xrpc subscriptions with current iteration of the spec

* Allow missing encoding for subscriptions

* track sequencedAt & eventType as well

* Resolve streaming type codes, streaming tests

* Fix interface for hooking into http server for streaming, fix stream routing

* Minor reorg how streaming is enabled in xrpc-server

* Server codegen for xrpc subscriptions, supporting types in xrpc-server

* fix up buffer overloading

* Rename data frame to message frame, rename subscription lex output to message

* Move body param first for streaming frames

* Tidy

* clean up evt types

* buff up tests

* missed merge conflict

* new schema

* blobs on subscriptions

* rm genned client subscription methods

* backfill limits

* testing subscription route & quick outbox bugfix

* fix up migration

* cascade on delete

* comments & naming

* fix dev env

* delete seqs on account deletion

* tidy

* fixing things up with db notify system for schemas

* fix duplicates in outbox

* tidy

* Break out of stream handler when socket closes

* fixing up some timing issues

* tidy

* terminate ws after using

* bump up timer on async reader

* fixing up NOTIFY in txns

* pr feedback

* pr bugfixes

* make order asc explicit

* bringing tests up to speed w atpagent

* bump up max listeners on sequencer

* increase timeouts for reading from generators

---------

+++ packages/pds/src/db/index.ts
@@ -1,11 +1,7 @@
 import assert from 'assert'
-import { Kysely, SqliteDialect, PostgresDialect, Migrator } from 'kysely'
+import { Kysely, SqliteDialect, PostgresDialect, Migrator, sql } from 'kysely'
 import SqliteDB from 'better-sqlite3'
-import {
-  Pool as PgPool,
-  PoolClient as PgPoolClient,
-  types as pgTypes,
-} from 'pg'
+import { Pool as PgPool, Client as PgClient, types as pgTypes } from 'pg'
 import EventEmitter from 'events'
 import TypedEmitter from 'typed-emitter'
 import DatabaseSchema, { DatabaseSchemaType } from './database-schema'
@@ -15,18 +11,26 @@ import { CtxMigrationProvider } from './migrations/provider'
 import { dbLogger as log } from '../logger'
 
 export class Database {
-  channels: Channels = {
-    repo_seq: new EventEmitter() as ChannelEmitter,
-  }
+  txChannelMsgs: ChannelMsg[] = []
+  channels: Channels
   migrator: Migrator
-  private channelClient: PgPoolClient | null = null
+  destroyed = false
 
-  constructor(public db: DatabaseSchema, public cfg: DialectConfig) {
+  private channelClient: PgClient | null = null
+
+  constructor(
+    public db: DatabaseSchema,
+    public cfg: DialectConfig,
+    channels?: Channels,
+  ) {
     this.migrator = new Migrator({
       db,
       migrationTableSchema: cfg.dialect === 'pg' ? cfg.schema : undefined,
       provider: new CtxMigrationProvider(migrations, cfg.dialect),
     })
+    this.channels = channels || {
+      repo_seq: new EventEmitter() as ChannelEmitter,
+    }
   }
 
   static sqlite(location: string): Database {
@@ -39,9 +43,8 @@ export class Database {
   }
 
   static postgres(opts: PgOptions): Database {
-    const { schema } = opts
-    const pool =
-      'pool' in opts ? opts.pool : new PgPool({ connectionString: opts.url })
+    const { schema, url } = opts
+    const pool = opts.pool ?? new PgPool({ connectionString: url })
 
     // Select count(*) and other pg bigints as js integer
     pgTypes.setTypeParser(pgTypes.builtins.INT8, (n) => parseInt(n, 10))
@@ -63,7 +66,7 @@ export class Database {
       dialect: new PostgresDialect({ pool }),
     })
 
-    return new Database(db, { dialect: 'pg', pool, schema })
+    return new Database(db, { dialect: 'pg', pool, schema, url })
   }
 
   static memory(): Database {
@@ -72,10 +75,14 @@ export class Database {
 
   async startListeningToChannels() {
     if (this.cfg.dialect !== 'pg') return
-    this.channelClient = await this.cfg.pool.connect()
-    await this.channelClient.query(`LISTEN repo_seq`)
+    if (this.channelClient) return
+    this.channelClient = new PgClient(this.cfg.url)
+    await this.channelClient.connect()
+    await this.channelClient.query(
+      `LISTEN ${this.getSchemaChannel('repo_seq')}`,
+    )
     this.channelClient.on('notification', (msg) => {
-      const channel = this.channels[msg.channel]
+      const channel = this.channels[this.normalizeSchemaChannel(msg.channel)]
       if (channel) {
         channel.emit('message')
       }
@@ -87,12 +94,49 @@ export class Database {
     })
   }
 
-  notify(channel: keyof Channels) {
+  async notify(channel: keyof Channels) {
     if (channel !== 'repo_seq') {
       throw new Error(`attempted sending on unavailable channel: ${channel}`)
     }
+    // hardcoded b/c of type system & we only have one msg type
+    const message: ChannelMsg = 'repo_seq'
+
+    // if in a sqlite tx, we buffer the notification until the tx successfully commits
+    if (this.isTransaction && this.dialect === 'sqlite') {
+      // no duplicate notifies in a tx per Postgres semantics
+      if (!this.txChannelMsgs.includes(message)) {
+        this.txChannelMsgs.push(message)
+      }
+    } else {
+      await this.sendChannelMsg(message)
+    }
+  }
+
+  private getSchemaChannel(channel: string) {
+    if (this.cfg.dialect === 'pg' && this.cfg.schema) {
+      return this.cfg.schema + '_' + channel
+    } else {
+      return channel
+    }
+  }
+
+  private normalizeSchemaChannel(schemaChannel: string): string {
+    if (this.cfg.dialect === 'pg' && this.cfg.schema) {
+      const prefix = this.cfg.schema + '_'
+      if (schemaChannel.startsWith(prefix)) {
+        return schemaChannel.slice(prefix.length)
+      } else {
+        return schemaChannel
+      }
+    } else {
+      return schemaChannel
+    }
+  }
+
+  private async sendChannelMsg(channel: ChannelMsg) {
     if (this.cfg.dialect === 'pg') {
-      this.cfg.pool.query(`NOTIFY ${channel}`)
+      const { ref } = this.db.dynamic
+      await sql`NOTIFY ${ref(this.getSchemaChannel(channel))}`.execute(this.db)
     } else {
       const emitter = this.channels[channel]
       if (emitter) {
@@ -102,10 +146,17 @@ export class Database {
   }
 
   async transaction<T>(fn: (db: Database) => Promise<T>): Promise<T> {
-    return await this.db.transaction().execute((txn) => {
-      const dbTxn = new Database(txn, this.cfg)
-      return fn(dbTxn)
+    let txMsgs: ChannelMsg[] = []
+    const res = await this.db.transaction().execute(async (txn) => {
+      const dbTxn = new Database(txn, this.cfg, this.channels)
+      const txRes = await fn(dbTxn)
+      txMsgs = dbTxn.txChannelMsgs
+      return txRes
     })
+    txMsgs.forEach((msg) => {
+      this.sendChannelMsg(msg)
+    })
+    return res
   }
 
   get schema(): string | undefined {
@@ -125,13 +176,12 @@ export class Database {
   }
 
   async close(): Promise<void> {
-    this.channelClient?.removeAllListeners()
-    this.channelClient?.release()
-    // @TODO investigate
-    // if (this.cfg.dialect === 'pg') {
-    //   await this.cfg.pool.end()
-    // }
+    if (this.destroyed) return
+    if (this.channelClient) {
+      await this.channelClient.end()
+    }
     await this.db.destroy()
+    this.destroyed = true
   }
 
   async migrateToOrThrow(migration: string) {
@@ -172,6 +222,7 @@ export type DialectConfig = PgConfig | SqliteConfig
 export type PgConfig = {
   dialect: 'pg'
   pool: PgPool
+  url: string
   schema?: string
 }
 
@@ -182,15 +233,19 @@ export type SqliteConfig = {
 // Can use with typeof to get types for partial queries
 export const dbType = new Kysely<DatabaseSchema>({ dialect: dummyDialect })
 
-type PgOptions =
-  | { url: string; schema?: string }
-  | { pool: PgPool; schema?: string }
+type PgOptions = {
+  url: string
+  pool?: PgPool
+  schema?: string
+}
 
 type ChannelEvents = {
   message: () => void
 }
 
 type ChannelEmitter = TypedEmitter<ChannelEvents>
+
+type ChannelMsg = 'repo_seq'
 
 type Channels = {
   repo_seq: ChannelEmitter
